feat(new tournament page): Initial setup
- Create the page to create a new tournament on the "/tournaments/new-
tournament" route. It should contain fields for entering the tournament
name, and the number of rounds, as well as a button that opens a model
where the user picks the class they want to create a turnament for.

- When a class is selected in the modal, it should display a list of
the classes' students, which lets the user pick what students should be
included in the tournament.