fix(points and pairings components)
- Make the number of matches completed be set in a function inside of a
useEffect hook. This solves the problem of having the number of matches
completed reset to 0 whenever switching between the Points and Pairings
components, which was causing the “Start next round” button to not
appear.

- Fix the issue of the wrong winner being displayed on the Points page
when the tournament is finished. The participantsData should be updated
when clicking the “Finish tournament” button.