feature(matches): Reset match
- Add a reset button to the Matches component that, when clicked, resets
the score of the match. If the score of a match is reset, the user
should not be able to click the “Next” button to go to the next match.
Nor should they be able to finish the tournament as long as the results
of one or more matches are missing. Once the tournament has finished, it
should no longer be possible to reset scores.

- Add a hover effect to the “Start Next Round”-button.

- Change the order of the links on the Profile page and in the Navbar to
be in the order you would use them to create a tournament, i.e. Classes
first, Tournaments second.

- Increase the size of the Points table at larger screen sizes.

- Decrease the top padding on the Profile page for screen sizes over
800px.